Add MatrixHandle class
